Move StashImageCardView to its own file
